Fixed C warning messages
Code (except for libfixmath) is now free of many optional GCC warnings.
The flags pointer-sign and conversion are suppressed because they are
problematic with Peter Danneggers 1wire code and libfixmath headers.

The following warning flags are enabled:
 * -Wall
 * -Wpedantic
 * -Wshadow
 * -Wcast-qual
 * -Wstrict-prototypes  (implies foo(void))
 * -Wmissing-prototypes (require prototypes for non-static functions)
 * -Wformat
 * -Wswitch-default
 * -Wcast-align
 * -Wbad-function-cast
 * -Winline
 * -Wundef
 * -Wnested-externs
 * -Wextra

+++ firmware/uart.h
@@ -66,7 +66,7 @@ LICENSE:
  *  @param  xtalCpu  system clock in Mhz, e.g. 4000000UL for 4Mhz          
  *  @param  baudRate baudrate in bps, e.g. 1200, 2400, 9600     
  */
-#define UART_BAUD_SELECT(baudRate,xtalCpu)  (((xtalCpu) + 8UL * (baudRate)) / (16UL * (baudRate)) -1UL)
+#define UART_BAUD_SELECT(baudRate,xtalCpu)  (unsigned int)(((xtalCpu) + 8UL * (baudRate)) / (16UL * (baudRate)) -1UL)
 
 /** @brief  UART Baudrate Expression for ATmega double speed mode
  *  @param  xtalCpu  system clock in Mhz, e.g. 4000000UL for 4Mhz
